# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Friday Food Club API — drops, orders, chefs, admin, subscriptions
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Create an account (BUYER or CHEF)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterInput' }
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in with email + password
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginInput' }
      responses:
        '200':
          description: Logged in
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
...
  /auth/area:
    patch:
      operationId: updateArea
      tags: [auth]
      summary: Set user location for the proximity feed
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AreaUpdate' }
      responses:
        '200':
          description: Area updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  area: { type: string }
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name:     { type: string }
        email:    { type: string }
        password: { type: string }
        role:     { type: string, enum: [BUYER] }
        area:     { type: string }
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:    { type: string }
        password: { type: string }
    AuthResponse:
      type: object
      properties:
        user:  { $ref: '#/components/schemas/SessionUser' }
        token: { type: string }
    SessionUser:
      type: object
      properties:
        id:                      { type: string }
        name:                    { type: string }
        email:                   { type: string }
        role:                    { type: string, enum: [BUYER, CHEF, ADMIN] }
        area:                    { type: string, nullable: true }
        chefId:                  { type: string, nullable: true }
        chefVerified:            { type: boolean, nullable: true }
        chefVerificationStatus:  { type: ["string", "null"], enum: [PENDING_REVIEW, VERIFIED, REJECTED, null] }
        chefRejectionReason:     { type: ["string", "null"] }
        isMember:                { type: boolean }
        walletBalance:           { type: number, nullable: true }
    AreaUpdate:
      type: object
      required: [area]
      properties:
        area: { type: string }
...